requirements.txt
numpy>=1.26
matplotlib>=3.8
numba>=0.59
Pillow>=10.0
